+++ config/graph/tsconfig.json
@@ -5,6 +5,10 @@
   "compilerOptions": {
     "outDir": "./dist",
     "rootDir": "./src",
+    "lib": ["ESNext"],
+    "module": "ESNext",
+    "moduleResolution": "Bundler",
+    "target": "ESNext",
     "declaration": true,
     "declarationMap": true,
     "sourceMap": true
