+++ packages/client/src/pure/pools/pools/pools.ts
@@ -1,4 +1,4 @@
-import { type getPoolsFromDB } from '../../../api/pools/pools/'
+import { type getPoolsFromDB } from '../../../api/pools/pools'
 import { EVM_APP_BASE_URL } from '../../../constants'
 import { parseArgs } from '../../../functions'
 import { type GetApiInputFromOutput } from '../../../types'
