+++ packages/extractor/tsconfig.json
@@ -7,6 +7,10 @@
     "baseUrl": ".",
     "outDir": "./dist",
     "rootDir": "./src",
+    "lib": ["ESNext"],
+    "module": "ESNext",
+    "moduleResolution": "Bundler",
+    "target": "ESNext",
     "declaration": true,
     "declarationMap": true,
     "sourceMap": true,
