+++ packages/graph-client/lib/incentives.ts
@@ -17,25 +17,19 @@ export interface Incentive {
 }
 
 export const getIncentives = async (): Promise<Incentive[]> =>
-  import('node-fetch').then(({ default: fetch }) =>
-    fetch('https://incentives.sushi.com/v0').then((data: any) => data.json()),
-  )
+  fetch('https://incentives.sushi.com/v0').then((data: any) => data.json())
 
 export const getIncentivesByPoolIds = async (
   poolIds: string[],
 ): Promise<Incentive[]> =>
-  import('node-fetch').then(({ default: fetch }) =>
-    fetch(`https://incentives.sushi.com/v0?poolIds=${poolIds.join(',')}`).then(
-      (data: any) => data.json(),
-    ),
+  fetch(`https://incentives.sushi.com/v0?poolIds=${poolIds.join(',')}`).then(
+    (data: any) => data.json(),
   )
 
 export const getIncentivesByPoolId = async (
   chainId: ChainId,
   address: string,
 ): Promise<Incentive[]> =>
-  import('node-fetch').then(({ default: fetch }) =>
-    fetch(`https://incentives.sushi.com/v0/${chainId}/${address}`).then(
-      (data: any) => data.json(),
-    ),
+  fetch(`https://incentives.sushi.com/v0/${chainId}/${address}`).then(
+    (data: any) => data.json(),
   )
