+++ packages/graph-client/package.json
@@ -41,32 +41,21 @@
     "preset": "@sushiswap/jest-config"
   },
   "dependencies": {
+    "@graphprotocol/client-cli": "3.0.0",
     "@graphprotocol/client-add-source-name": "2.0.0",
     "@graphprotocol/client-auto-pagination": "2.0.0",
     "@graphprotocol/client-auto-type-merging": "2.0.0",
-    "@graphql-mesh/cache-localforage": "0.94.1",
-    "@graphql-mesh/cross-helpers": "0.4.0",
-    "@graphql-mesh/graphql": "0.94.2",
-    "@graphql-mesh/http": "0.94.4",
-    "@graphql-mesh/merger-stitching": "0.94.1",
-    "@graphql-mesh/runtime": "0.94.1",
-    "@graphql-mesh/store": "0.94.1",
-    "@graphql-mesh/transform-filter-schema": "0.94.1",
-    "@graphql-mesh/transform-prefix": "0.94.2",
-    "@graphql-mesh/transform-prune": "0.94.1",
-    "@graphql-mesh/transform-rename": "0.94.2",
-    "@graphql-mesh/transform-type-merging": "0.94.1",
-    "@graphql-mesh/types": "0.94.1",
-    "@graphql-mesh/utils": "0.94.1",
-    "@graphql-tools/merge": "9.0.0",
+    "@graphql-mesh/transform-filter-schema": "0.96.6",
+    "@graphql-mesh/transform-prefix": "0.96.6",
+    "@graphql-mesh/transform-prune": "0.96.6",
+    "@graphql-mesh/transform-rename": "0.96.7",
+    "@graphql-mesh/transform-type-merging": "0.96.6",
+    "@graphql-tools/merge": "9.0.3",
     "@sushiswap/graph-config": "workspace:*",
     "@sushiswap/wagmi-config": "workspace:*",
-    "date-fns": "2.30.0",
-    "node-fetch": "3.3.0",
-    "tslib": "2.5.0"
+    "date-fns": "2.30.0"
   },
   "devDependencies": {
-    "@graphprotocol/client-cli": "3.0.0",
     "@sushiswap/jest-config": "workspace:*",
     "@sushiswap/typescript-config": "workspace:*",
     "@types/jest": "29.5.5",
