+++ packages/sushi/package.json
@@ -13,6 +13,7 @@
   },
   "license": "UNLICENSED",
   "author": "Matthew Lilley <hello@matthewLilley.com>",
+  "type": "module",
   "contributors": [
     "Ilya Lyalin",
     "Ola Stenberg",
@@ -185,7 +186,7 @@
     "check": "tsc --pretty --noEmit",
     "clean": "rm -rf .turbo && rm -rf node_modules && rm -rf dist",
     "dev": "tsc -w",
-    "generate": "ts-node --skipProject ./scripts/generate.ts",
+    "generate": "npx tsx ./scripts/generate.ts",
     "prepublishOnly": "pnpm build",
     "test": "vitest run -c ./test/vitest.config.ts",
     "test:debug": "vitest --inspect-brk --no-threads run -c ./test/vitest.config.ts",
